extract functions from app.js to make things cleaner
